Hu 3 4 9 (#2)
* *FIX: Fix conflicts CSS
*ADD: Added session control (LogIn, LogOut)
*ADD: Added HU_3; Search game
*ADD: Added HU_4; View game
*ADD: Added HU_9; As a player I want to post a message
*ADD: Added dependency of mysql connector java in maven (pom.xml)

The unregistered user will not be able to post a message.
The header interface will be modified by exchanging LogIn or LogOut
depending on whether the user is registered or not.

* *UPDATE: View of search games

Now the result of search games can be view together in one table

* * Refactoring code

Now from the game search window you can also
see the platform of each game